third_party/WebKit/LayoutTests/accessibility/aria-help.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<div id="button" tabindex=0 role="button" aria-label="title" aria-help="click here">
button
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-help attribute works as expected.");

    if (window.accessibilityController) {

        document.getElementById("button").focus();
        var button = accessibilityController.focusedElement;
        shouldBe("button.deprecatedHelpText", "'AXHelp: click here'");
    }

</script>

</body>
</html>
